Memoize spanmetrics sanitizelabelname, move to better location
